requirements.txt
cdsapi
fitter
geopandas
IPython
mango-optimizer
mapply
matplotlib
miceforest
mplcyberpunk
numpy
pandarallel
pandas
plotly
plotnine
pygeohash
requests
scikit-learn
scikit-plot
scipy
seaborn
statsmodels
termcolor
tqdm
xarray
xgboost
yellowbrick